backend/pyproject.toml
[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "Context-aware RAG chatbot backend for Physical AI textbook"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "qdrant-client>=1.7.0",
    "pydantic>=2.0.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
